Experiment: make schema validation explicit.
The state utils confusingly allowed for schema validation to be added
to reducers in a number of ways:
- Passing a third parameter to createReducer
- Adding a `schema` property to a plain reducer function
- Wrapping a plain reducer function with withSchemaValidation

The first two, in particular, were implicit and hard to determine at
compile time, leading to some strong coupling in the state utils, and
requiring validation code to always be present, even if none of the
reducers use it.

This change makes schema validation explicit, by one of two methods:
- Using createReducerWithValidation instead of createReducer
- Wrapping a plain reducer function with withSchemaValidation

This makes it possible to tree-shake effectively at compile time, should
we ever drop enough schemas, and it makes state utils less coupled, by
e.g. making combineReducers no longer depend on validation code.